requirements.txt
fastapi
uvicorn
pillow
pytesseract
ultralytics
torch
requests
